requirements.txt
alembic==1.7.1
certifi==2021.5.30
charset-normalizer==2.0.4
ci-info==0.2.0
click==8.0.1
colorama==0.4.4
configobj==5.0.6
configparser==5.0.2
dataclasses==0.8
decorator==4.4.2
dnspython==2.1.0
email-validator==1.1.3
etelemetry==0.2.2
filelock==3.0.12
fitz==0.0.1.dev2
Flask==2.0.1
Flask-Admin==1.5.8
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
future==0.18.2
greenlet==1.1.1
httplib2==0.19.1
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
Mako==1.1.5
MarkupSafe==2.0.1
networkx==2.5.1
nibabel==3.2.1
nipype==1.6.1
numpy==1.19.5
packaging==21.0
pandas==1.1.5
pathlib==1.0.1
Pillow==8.3.2
prov==2.0.0
pydot==1.4.2
PyMuPDF==1.18.17
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
pyxnat==1.4
rdflib==5.0.0
requests==2.26.0
scipy==1.5.4
simplejson==3.17.5
six==1.16.0
SQLAlchemy==1.4.23
traits==6.2.0
typing-extensions==3.10.0.2
urllib3==1.26.6
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.5.0
